pr improvements
